requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli @ file:///croot/brotli-split_1714483155106/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1707229174982/work/certifi
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
datasets==2.19.1
diffusers==0.28.0
dill==0.3.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
errant==3.0.0
filelock @ file:///croot/filelock_1700591183607/work
fire==0.6.0
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gdown==5.2.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@23425cd2e98a919384cab6156af8adf1c9d0639a
huggingface-hub==0.23.1
idna==2.10
imhist==0.0.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
imWatermark==0.0.2
invisible-watermark==0.2.0
jieba==0.42.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
langcodes==3.4.0
language-tool-python==2.8
language_data==1.2.0
Levenshtein==0.25.1
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx @ file:///croot/networkx_1690561992265/work
nlprule==0.6.4
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
Pillow==9.0.1
portalocker==2.8.2
preshed==3.0.9
protobuf==5.27.0
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.25.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torchattacks==3.5.1
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
typer==0.9.4
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
